Fix XML parsing bug in harnessgen
We have translated the XML file from UTF-8 to CP1252, but the charset
declaration on top of the file hasn't been updated, and the XML parser gets
confused thinking it's parsing a UTF-8 document. To fix this, we manually parse
the file using the correct charset before feeding it to the parser.