Untrack more files -- be careful when adding files, see comment for more
We all have local modifications and temporary files in our tree that shouldn't
be uploaded to the repo. .gitignore covers many common cases of such files, but
we haven't thought of everything, so please check what you're commiting. Follow
these simple rules, and we can minimize the amount of garbage that makes it to
the repo:

- NEVER do "git commit -a"
- NEVER do "git add -A"
- AVOID using wildcards (asterisks) in "git add"
- To only commit part of a file, use "git add -i"
- Try using a GUI for building your commits, e.g. the default git GUI:
  "git gui"